Added possiility to view saved galaxies and edit them, while still able to create new galaxies.

+++ src/main/java/com/starfarers/dao/BaseDao.java
@@ -12,7 +12,7 @@ public interface BaseDao<T> {
 
 	public void remove(T entity);
 
-	public T find(Long id);
+	public T find(Integer id);
 
 	public List<T> findAll();
